requirements-dev.txt
# Development & Testing Requirements
# Install with: pip install -r requirements-dev.txt

# ============================================
# Include all production requirements
# ============================================
-r requirements.txt

# ============================================
# Testing Framework
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
pytest-html>=4.0.0

# ============================================
# UI Testing
# ============================================
playwright>=1.40.0
pytest-playwright>=0.4.0
# selenium>=4.15.0  # Alternative to Playwright

# ============================================
# HTTP Testing
# ============================================
requests-mock>=1.11.0
responses>=0.23.0
httpretty>=1.1.4

# ============================================
# Test Data Generation
# ============================================
factory-boy>=3.3.0
faker>=19.0.0

# ============================================
# Code Quality Tools
# ============================================
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# ============================================
# Performance Testing
# ============================================
memory-profiler>=0.61.0
# locust>=2.17.0  # Load testing - uncomment if needed

# ============================================
# Parallel Testing
# ============================================
pytest-xdist>=3.3.0  # Run tests in parallel

# ============================================
# Test Reporting
# ============================================
pytest-json-report>=1.5.0
# allure-pytest>=2.13.0  # Advanced reporting - uncomment if needed